Refactor validation test to be table driven.
